Move QRCode class to own file
